+++ src/main/java/com/spotpobre/backend/infrastructure/security/config/SecurityConfig.java
@@ -72,6 +72,8 @@ public class SecurityConfig {
                         .requestMatchers(HttpMethod.PATCH, "/api/v1/playlists/**").authenticated()
                         .requestMatchers(HttpMethod.DELETE, "/api/v1/playlists/**").authenticated()
                         .requestMatchers(HttpMethod.GET, "/api/v1/playlists/**").authenticated()
+                       .requestMatchers(HttpMethod.GET, "/api/v1/artists").authenticated()
+                       .requestMatchers(HttpMethod.GET, "/api/v1/artists/*/albums").authenticated()
                         .requestMatchers("/api/v1/me/playlists").authenticated()
                         .requestMatchers(HttpMethod.GET, "/api/v1/songs/**").authenticated()
                         .requestMatchers(HttpMethod.GET, "/api/v1/artists/**").authenticated()